refactor: CardNickNameForm, TextInput index 모듈 추가
