+++ src/components/common/index.tsx
@@ -1,0 +1,2 @@
+export { default as CardNickNameForm } from './CardNickNameForm';
+export { default as TextInput } from './TextInput';
